オブジェクトの構造を変更し、アイテム取得とアイテム名表示に関してのコライダーの判定を２分化して別に作動するように修正。各スクリプトを修正。サマリー追加。 リファクタリングして MVP パターン化できるようにクラス名を Model、View 付きに変更。デバッグ完了。
