+++ Assets/Scripts/ItemDetailModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b50235d926765234aa88afd73cccaad0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
